Clean useless attributes on User
First part which only ignore columns (drop will be in another commit).

Some notes:

* `context` used to store (sometimes) the siret of the user, this commit
  use (if there is a token) the siret associated to the authorization
  request (not perfect, neither the old solution tbh)
