+++ app/interactors/datapass_webhook/find_or_create_user.rb
@@ -19,8 +19,7 @@ class DatapassWebhook::FindOrCreateUser < ApplicationInteractor
       'first_name' => user_attributes['given_name'],
       'last_name' => user_attributes['family_name'],
       'phone_number' => user_attributes['phone_number'],
-      'oauth_api_gouv_id' => user_attributes['uid'],
-      'context' => datapass_attributes['siret']
+      'oauth_api_gouv_id' => user_attributes['uid']
     }
   end
