+++ app/interactors/user/find_or_create_new_owner.rb
@@ -1,7 +1,6 @@
 class User::FindOrCreateNewOwner < ApplicationInteractor
   def call
     context.target_user = User.find_or_initialize_by_email(context.target_user_email)
-    context.target_user.context = context.current_owner.context
     fail!('invalid_email', 'warn', errors) unless context.target_user.save
   end
