+++ app/views/user_mailer/transfer_ownership.html.erb
@@ -7,17 +7,20 @@
     <p>Bonjour et bienvenue,</p>
 
     <p><%= @old_owner.email %> propose de vous transférer la responsabilité du compte utilisateur API Entreprise de votre organisation. Ce compte vous permettra d'accéder aux jetons (tokens) API Entreprise, à leurs statistiques d'utilisation et à l'interface de téléchargement des attestations si vos jetons sont pourvus de ces droits.</p>
-    
+
     <% if @new_owner.oauth_api_gouv_id.nil? %>
       <p>
       La poursuite des étapes suivantes entrainera le transfert du compte et toute connexion à l'espace utilisateur de votre part vaudra acceptation de nos <a href="https://entreprise.api.gouv.fr/cgu">CGU</a>.
-      </p>  
+      </p>
       <p>
       Pour accéder à l'espace utilisateur API Entreprise, vous devez au préalable vous créer un <%= link_to 'compte DataPass', @datapass_signup_url %>. Il est impératif de renseigner les informations suivantes afin d'éviter des erreurs de réconciliation à votre première connexion :
       </p>
       <ul>
         <li>adresse email : <%= @new_owner.email %></li>
-        <li>siret de votre organisation : <%= @new_owner.context %></li>
+
+        <% if @new_owner.siret %>
+          <li>siret de l'organisation: <%= @new_owner.siret %></li>
+        <% end %>
       </ul>
     <% else %>
       <p>
@@ -26,9 +29,9 @@
     <% end %>
     <p>Le transfert définitif de la responsabilité du compte nécessite une validation manuelle de nos services et peut prendre quelques jours.</p>
     <p>Vous pourrez ensuite accéder à l'<%= link_to 'espace utilisateur API Entreprise de votre organisation', @login_url %> en utilisant vos identifiants DataPass.</p>
-    
+
     <p>Vous ne souhaitez pas prendre la responsabilité du compte actuellement détenu par <%= @old_owner.email %> ou avez des interrogations ? Vous pouvez nous écrire par retour de mail.
-    
+
     <p>Cordialement,</p>
     <p>L'équipe API Entreprise</p>
   </body>
